+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+## 🌹 Release 89.0.0 - 2026-03-02
+
 ### Changed
 * **BREAKING** Legend: Legend en Legend Item fuseren tot één component Legend ([#3490](https://github.com/dso-toolkit/dso-toolkit/issues/3490))
 * Ozon Content: i-tje verwijderen bij verwijzing naar begrip (IntRef naar `<Begrip>`) ([#3562](https://github.com/dso-toolkit/dso-toolkit/issues/3562))